Create common module for security
